prob/requirements.txt
django
django-json-field
jupyter ipython django-extensions